refactor: rename replay indexer to counter
