+++ src/counter/counter_worker.ts
